// packages/virelia-web/src/components/RadioGroup/types.ts
import type {
  BaseRadioGroupProps,
  BaseRadioOption,
  Orientation,
} from '@romanbakurov/virelia-types';

export interface RadioOption extends BaseRadioOption {
  label: string;
}

export interface RadioGroupProps extends Omit<BaseRadioGroupProps, 'options'> {
  label?: string;
  name: string;
  options: RadioOption[];
  error?: string;
  orientation: Orientation;
  className?: string;
}